requirements.txt
yahooquery==2.3.7
CurrencyConverter==0.17.23
gpt-researcher==0.7.0
streamlit
numpy==1.26.4
pandas==2.0.3
requests
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
llama_index==0.10.52
pinecone-client==3.2.2
pinecone==4.0.0
schedule==1.2.2
docx2txt==0.8
torch==2.3.1
torchvision==0.18.1
transformers==4.42.3
python-pptx==0.6.23
Pillow==10.4.0
llama-index-vector-stores-pinecone==0.1.7
llama-index-readers-google==0.2.9
anthropic==0.28.1
llama-index-llms-anthropic==0.1.15
pytz==2024.1
llama-index-embeddings-huggingface==0.2.2
openpyxl==3.1.5
